<DE2 Database>
Function modified:
- initPlaylist() now needs a char* listname as an argument.
Function added:
- loadListsFromSD() : loads all PLSTxx.TXT file from sdcard, writes all
list & its ID info to LISTS.TXT, and adds all playlist to database.
- openFileFromSD() : opens a file and handles all error
making this as functions since many functions need to open file and
handle errors (reusability).
- getListNameFromListFile() : given a PLST1.TXT or any playlist text
file, read the first line of that file to retrieve the list name. use
this to store playlist to DB.

Note:
Playlist is used to be stored as LISTxx.TXT, but I changed to
PLSTxx.TXT. This way, LISTS.TXT (text file that consists playlists and
their IDs) will not be considered as a playlist.
Hope it make sense =]