Add setting of volumes for individual sounds
I've added a static volume change, which will rewrite all of the sound
buffer so that the volume multiplication doesn't need to happen whenever
the sound is read from memory.

I've also added a dynamic volume change, which allows the sound to
maintain it's original wave form so that the sound volume can be brought
all the way to 0 and then back to 1 without losing any data.  This takes
up more processing, but is the expected behaviour.