Add fading in and out of songs
Right now this is a pretty quick and dirty implementation.  If the sound
is within the in fade time our the out fade time, the volume will be
updated after every audio interrupt.  Using a float for the volume
destroys the quality of the sound...this may need to be changed soon.