Add setPitch functions that use PhaseVocoder
These functions allow the pitch of a sound wave to be set with varying
quality.  With a quality of 0 the sound's tempo will be changed when the
pitch is changed.  Above a quality of 0 the PhaseVocoder will be used to
allow for pitch change without an effect on the sound duration.  Right now
this is quite slow, a quality of 1 should be used.