<DE2> highlightSongInList Implemented
on this commit:

highlightSongInList(int list_id, int song_id) is implemented. This
function will highlight a song with given song_id, it will check if that
song exists in the first 14th song in that list. If it exists in the
order > 14, then it will not do anything.

Note that when calling this function, you have to have all songs in that
playlist drawn on the panel (current panel has to be 2) otherwise it
will not do anything.